// components/main/agreement/types.tsx
export type agreement = {
  id: string;
  subsidiaryId: string;
  bourseCodeType: number;
  name: string;
  description: string;
  context: string;
  defaultFile: any;
  isBourseCodeRequired: boolean;
  isRequired: boolean;
  isActive: boolean;
  isDeleted: boolean;
  code: number;
  createDateTime: string;
  updateDateTime: string;
};